Add retrievePetition to submitCode pipeline
Adds a new step to the `submitCode` pipeline so the petition information
is returned to the frontend via the API in the same way we return that
information in the `verifySignature` pipeline. This fixes the issue
where a message is displayed to the user after submitting the code
rather than the petition title.